+++ src/apps/curve/contracts/index.ts
@@ -4,26 +4,30 @@ import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
 import { ContractFactory } from '~contract/contracts';
 import { Network } from '~types/network.interface';
 
-import { CurveAddressResolver__factory } from './ethers';
-import { CurveChildLiquidityGauge__factory } from './ethers';
-import { CurveChildLiquidityGaugeFactory__factory } from './ethers';
-import { CurveController__factory } from './ethers';
-import { CurveCryptoFactory__factory } from './ethers';
-import { CurveCryptoRegistry__factory } from './ethers';
-import { CurveDoubleGauge__factory } from './ethers';
-import { CurveFactory__factory } from './ethers';
-import { CurveFactoryPoolV2__factory } from './ethers';
-import { CurveGauge__factory } from './ethers';
-import { CurveGaugeV2__factory } from './ethers';
-import { CurveNGauge__factory } from './ethers';
-import { CurvePool__factory } from './ethers';
-import { CurvePoolLegacy__factory } from './ethers';
-import { CurveRewardsOnlyGauge__factory } from './ethers';
-import { CurveStableFactory__factory } from './ethers';
-import { CurveStableRegistry__factory } from './ethers';
-import { CurveVestingEscrow__factory } from './ethers';
-import { CurveVotingEscrow__factory } from './ethers';
-import { CurveVotingEscrowReward__factory } from './ethers';
+import {
+  CurveAddressResolver__factory,
+  CurveChildLiquidityGauge__factory,
+  CurveChildLiquidityGaugeFactory__factory,
+  CurveController__factory,
+  CurveCryptoFactory__factory,
+  CurveCryptoRegistry__factory,
+  CurveDoubleGauge__factory,
+  CurveFactory__factory,
+  CurveFactoryPoolV2__factory,
+  CurveGauge__factory,
+  CurveGaugeV2__factory,
+  CurveNGauge__factory,
+  CurvePool__factory,
+  CurvePoolLegacy__factory,
+  CurveRewardsOnlyGauge__factory,
+  CurveStableFactory__factory,
+  CurveStableRegistry__factory,
+  CurveTricryptoFactory__factory,
+  CurveTricryptoPool__factory,
+  CurveVestingEscrow__factory,
+  CurveVotingEscrow__factory,
+  CurveVotingEscrowReward__factory,
+} from './ethers';
 
 // eslint-disable-next-line
 type ContractOpts = { address: string; network: Network };
@@ -85,6 +89,12 @@ export class CurveContractFactory extends ContractFactory {
   curveStableRegistry({ address, network }: ContractOpts) {
     return CurveStableRegistry__factory.connect(address, this.appToolkit.getNetworkProvider(network));
   }
+  curveTricryptoFactory({ address, network }: ContractOpts) {
+    return CurveTricryptoFactory__factory.connect(address, this.appToolkit.getNetworkProvider(network));
+  }
+  curveTricryptoPool({ address, network }: ContractOpts) {
+    return CurveTricryptoPool__factory.connect(address, this.appToolkit.getNetworkProvider(network));
+  }
   curveVestingEscrow({ address, network }: ContractOpts) {
     return CurveVestingEscrow__factory.connect(address, this.appToolkit.getNetworkProvider(network));
   }
@@ -113,6 +123,8 @@ export type { CurvePoolLegacy } from './ethers';
 export type { CurveRewardsOnlyGauge } from './ethers';
 export type { CurveStableFactory } from './ethers';
 export type { CurveStableRegistry } from './ethers';
+export type { CurveTricryptoFactory } from './ethers';
+export type { CurveTricryptoPool } from './ethers';
 export type { CurveVestingEscrow } from './ethers';
 export type { CurveVotingEscrow } from './ethers';
 export type { CurveVotingEscrowReward } from './ethers';