+++ src/apps/polynomial/optimism/polynomial.smart-wallet.contract-position-fetcher.ts
@@ -25,6 +25,8 @@ export class OptimismPolynomialSmartWalletContractPositionFetcher extends Contra
   supportedTokens = [
     '0x8c6f28f2f1a3c87f0f938b96d27520d9751ec8d9', //sUSD
     '0x7f5c764cbc14f9669b88837ca1490cca17c31607', //USDC
+    '0x94b008aa00579c1307b0ef2c499ad98a8ce58e58', //USDT
+    '0xda10009cbd5d07dd0cecc66161fc93d7c9000da1', //DAI
     '0x4200000000000000000000000000000000000042'  //OP
   ];
   groupLabel = 'Smart Wallet';